Moved types to Scripting & GameCenter
